+++ CHANGELOG.md
@@ -37,6 +37,7 @@ use the TRACK menu to select the specific track to record from
   - When you press a pad in a new track, a popup will appear asking you to confirm the type of clip you wish to create. By default it will select the last clip type you created as the clip type to create. It will tell you on the display what that clip type is. 
     - If you just a tap a pad quickly to create a new clip, it will create that new clip using the last clip type.
     - If you press and hold a pad, you can choose a different type by pressing one of the `BLINKING` clip type buttons (e.g. `SYNTH`, `KIT`, `MIDI`, `CV` or `CROSS SCREEN` (for Audio Clips)). If you let go of the pad without selecting a different type, it will create the clip using the last selected clip type.
+    - If you press `BACK` before releasing a pad or selecting a clip type, it will cancel the clip creation.    
     - These changes only apply to `SONG GRID VIEW` and NOT `SONG ROW VIEW`
 - Updated the `SONG GRID VIEW` shortcut of `HOLDING PAD FOR THE CLIP` + `CROSS SCREEN BUTTON` to convert an Empty `INSTRUMENT CLIP` to an `AUDIO CLIP`.
   - This replaces the previous shortcut of `HOLDING PAD FOR THE CLIP` + `SELECT ENCODER`.